Expand AEAD limits to consider multi-user security.
In a typical multi-user setting, the adversary is assumed to perform some
(massive) amount of offline work to break the integrity of a single random
connection. Each user is represented as a unique (key, nonce) pair. In
considering the same threat model, we must treat each key resulting from
a KeyUpdate event as a unique user. As a result, any bounds implied from
the multi-user setting translate (minimally) to bounds for multiple keys
used in the course of a single connection. Note that this is not perfect,
as the total number of users in a standard multi-user setting is far larger
than the number of keys used in a single connection. However, absent realistic
"per-connection" bounds, this is the best we can do. (This is a gap in
the literature.)

Hoang et al. present tight multi-user security bounds for randomized AES-GCM
(as is used in TLS 1.3 and QUIC), so we can take advantage of those for
per-connection limits. In contrast, there are no multi-user security bounds
for AEAD_CHACHA20_POLY1305 or AEAD_AES_128_CCM, so we must use the
single-user bounds in their stead.